+++ spec/features/transfers/users_can_create_a_transfer_project_spec.rb
@@ -58,6 +58,10 @@ RSpec.feature "Users can create new conversion projects" do
       choose "No"
     end
 
+    within("#financial-safeguarding-governance-issues") do
+      choose "No"
+    end
+
     fill_in "Handover comments", with: "This is a handover note."
 
     within("#provisional-transfer-date") do